Move as_user helper into dedicated module
This change introduces a helper module, 'common', to the integration
test suite and moves the as_user() helper function into it. In so doing
we pave the way for using this functionality in other tests.
